Kaplan-Meier Plot for Continuous or Categorical Covariates

This function generates a Kaplan-Meier plot for time-to-event data, stratified either by quartiles of a continuous covariate or by levels of a categorical covariate.

kmplot(
  data,
  time_var,
  event_var,
  covariate,
  covariate_type = c("continuous", "categorical"),
  covariate_labels = NULL,
  covariate_label = NULL,
  time_label = NULL,
  survival_label = NULL,
  colors = NULL,
  time_break = 1,
  plot_ci = FALSE,
  show_pval = FALSE,
  show_fit = TRUE,
  cumulative = FALSE,
  return_plot = TRUE,
  show_summary_box = TRUE,
  summary_box_position = c("topright", "topleft", "bottomright", "bottomleft")
)

Arguments

data

A data.frame containing the input data.

time_var

Character. Name of the time-to-event variable.

event_var

Character. Name of the event indicator variable (0 = censored, 1 = event).

covariate

Character. Name of the covariate to stratify by.

covariate_type

Character. One of "continuous" or "categorical".

covariate_labels

Optional named vector to rename levels of the covariate.

covariate_label

Optional character. Label for the covariate (used in legend and risk table).

time_label

Optional character. Label for the x-axis.

survival_label

Optional character. Label for the y-axis.

colors

Optional character vector of colors for the groups.

time_break

Numeric. Interval for x-axis breaks (default is 1).

plot_ci

Logical. Whether to show confidence intervals (default = FALSE).

show_pval

Logical. Whether to display the p-value (default = FALSE).

show_fit

Logical. Whether to print the survfit object (default = TRUE).

cumulative

Logical. Whether to plot cumulative incidence instead of survival (default = FALSE).

return_plot

Logical. Whether to return the ggplot object (default = TRUE).

Value

A Kaplan-Meier plot (ggplot2 object) with optional risk table.

Examples

library(survival)
library(survminer)
#> Loading required package: ggplot2
#> Loading required package: ggpubr
#> 
#> Attaching package: ‘survminer’
#> The following object is masked from ‘package:survival’:
#> 
#>     myeloma
data(lung)
#> Warning: data set ‘lung’ not found

# Example 1: Categorical covariate (sex)
kmplot(data = lung,
time_var = "time",
    event_var = "status",
    covariate = "sex",
    covariate_type = "categorical",
    covariate_labels = c("1" = "Male", "2" = "Female"),
    covariate_label = "Sex",
    time_label = "Time (days)",
    survival_label = "Survival Probability",
    time_break = 100,
    show_pval = TRUE)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
#> Call: survfit(formula = Surv(time, event) ~ lbl, data = data)
#> 
#>              n events median 0.95LCL 0.95UCL
#> lbl=Female  90     53    426     348     550
#> lbl=Male   138    112    270     212     310
#> Error in exposr_palette(min(n_groups, 8)): could not find function "exposr_palette"
# Example 2: Continuous covariate (age)
kmplot(data = lung,
    time_var = "time",
    event_var = "status",
    covariate = "age",
    covariate_type = "continuous",
    covariate_label = "Age Quartiles",
    time_label = "Time (days)",
    survival_label = "Survival Probability",
    time_break = 100,
    show_pval = TRUE)
#> Call: survfit(formula = Surv(time, event) ~ lbl, data = data)
#> 
#>             n events median 0.95LCL 0.95UCL
#> lbl=Q1: 52 58     39    337     239     457
#> lbl=Q2: 60 59     41    348     245     574
#> lbl=Q3: 67 55     39    329     285     477
#> lbl=Q4: 73 56     46    283     222     361
#> Error in exposr_palette(min(n_groups, 8)): could not find function "exposr_palette"